toolchain: Remove the support for building GDB
The multiarch builds of GDB are generally available in all major Linux
distributions since 2022 (with early adopters as far back as in 2012
and latecomers in 2024). We have stopped building per-target GDBs by
default in 2023, seeing little benefit in doing that.

This change removes that obsolete feature completely. If there is a
need to provide a custom build of a multiarch GDB that targets all
architectures supported by HelenOS, it can be provided as a separate
script.